Add registry of objects created by migrations
Every time the object is created by a migration,
we'll register it in the special table `pg_extra`
with oid, classid and version. It prevents unwanted objects
from being added to the database schema.

The table should be created and dropped along with
the Rails `schema_migrations` to make the database
ready for being migrated. This is done by the gem's railtie.