Add callbacks to operations
The only callback we use is the `after_initialize`
which will be invoked after the end of the object
initialization with explicit arguments.

It enables autogeneration of missed values
(like the name of the object) which can be extracted
from the other assigned attributes.

Add `after_initialize` callback to operation

In some cases the missed arguments (like the name of the index etc.)
can be generated from the rest of the operation parameters.
We cannot define these values as defaults because the later are
static, while we need to take other arguments into consideration.

To do this idiomatically, use the `after_initialize` callback.
The callback will be invoked at the very end of the initialization
when explicit arguments are assigned and a block containing new
definitions is invoked.

  attribute :name, :string
  attribute :table, :string
  attribute :column, :string

  validates :table, :column, presence: true

  after_initialize do
    self.name ||= "#{table}_by_{column}_idx"
  end