Register new operations
New definitions should be reflected in the corresponding
Rails classes and modules via Railtie.

We make registration automatically every time when
a corresponding method is added to the operation definition:
- to_sql (we know how to execute the operation)
- invert (we know how to invert the operation)

The registration is also (re-)done after the invocation
of the `ruby_params` method because it changes
the signature of the operation method.