Add attributes to operations
The class method helper `attribute` would define
the attribute of the operation.

Attributes can be sent as a hash argument and
be coerced using various serializers.

In addition to the `ActiveModel::Attributes` API,
we enable setting aliases for the attributes.
They are applied during the instantiation only:
you can use the alias instead of the attribute name:
It is useful when you need to provide singular version:

  attribute :columns, :pg_extra_array_of_strings, aliases: %i[column]

With the definition above you can send either

  { column: :foo, ... }

or:

  { columns: %w[foo bar], ... }

Deserialized values in any case will be assigned to
the `columns` attribute (the `column` is just the key, no more).

The operatoion definition can be converted back to
the list of its (serialized) attributes using
the `#attributes` (aliased as `#to_h`) method call.