Add order to operations
The order enables sorting operations fetched from the db.

The order of operations of different kinds will be defined
by the order of their loading.

The order of operations of the same kind can be specified using
the operator-specific method `<=>`. By default,
they are ordered by names (presumably unique in PostgreSQL).
You can overload this order for every kind of operation.