Add inversion of operations
The operation can be reverted. To do this you should
specify the `invert` instance method which return
a definition of the inverted operation.

The method can also return `nil` which means that
inversion is not needed (can be skipped). This is necessary
because some operations like validation of foreign keys
or check constraints is unidirectional. When the corresponding
migration is reverted you should do nothing.

If the operation cannot be reverted (for example, you
cannot revert adding values to the enumerable type,
neither you can drop and recreate used type), just
call the `irreversible!(*messages)` from the `invert` method.
This would raise the `PGExtra::IrreversibleMigration`
carrying a verbose description of the problem.