Add validation of operations
The standard ActiveModel validations are added to
the operation definition. Use method `validates` and `validate`
to check if the operation is prepared correctly.

Notice, that we use validation in 2 scopes
- before the operation's execution
- when building an inversion

In the second case we'll re-raise validation error
as `PGExtra::IrreversibleMigration` carrying detailed
explanation why inverted operation cannot be exectuted.
(As a rule this is caused when the `DROP`-ping some object
carries not enough arguments to re-`CREATE` it back).