Add generator-related attribute options
Later we will define the rails generators for different operations
(mostly for `add_` and `create_` ones).

Operation definitions has already kept all the necessary
settings that can be reused by generators. The only missed
part is the description of attributes for the help.

To support this feature we add the additional `:desc`
key to the operation's `.attribute` helper.