Add operations with functions
We support function-related DDL in a different manner
when f(x) does. Instead of usage of unsafe SQL snippet
we ask for the SQL body only. The rest of the definition
should be made via specific options of the `create_function`
method call.

The following definitions are supported:
- create_function
- rename_function
- change_function
- drop_function

Notice, that the `change_function` is different from the f(x)
method `update_function` which drops'n'creates a function.
Instead we use `ALTER FUNCTION` SQL command to change some
parameters only.